feat: add typescript
