+++ src/app.ts
@@ -1,0 +1,34 @@
+// const express = require('express');
+// const swaggerUI = require('swagger-ui-express');
+// const path = require('path');
+// const YAML = require('yamljs');
+
+import express from 'express';
+import swaggerUI from 'swagger-ui-express';
+import path from 'path';
+import YAML from 'yamljs';
+import userRouter from './resources/users/user.router';
+import boardsRouter from './resources/boards/board.router';
+// const userRouter = require('./resources/users/user.router');
+// const boardsRouter = require('./resources/boards/board.router');
+
+
+const app = express();
+const swaggerDocument = YAML.load(path.join(__dirname, '../doc/api.yaml'));
+
+app.use(express.json());
+
+app.use('/doc', swaggerUI.serve, swaggerUI.setup(swaggerDocument));
+
+app.use('/', (req, res, next) => {
+  if (req.originalUrl === '/') {
+    res.send('Service is running!');
+    return;
+  }
+  next();
+});
+
+app.use('/users', userRouter);
+app.use('/boards', boardsRouter);
+
+module.exports = app;
